almost done doing kml
